+++ src/Plugin/Alipay/Fund/AccountQueryPlugin.php
@@ -16,7 +16,7 @@ class AccountQueryPlugin implements PluginInterface
 {
     public function assembly(Rocket $rocket, Closure $next): Rocket
     {
-        Logger::info('[alipay][AccountQueryPlugin] 插件开始装载', ['rocket' => $rocket]);
+        Logger::debug('[alipay][AccountQueryPlugin] 插件开始装载', ['rocket' => $rocket]);
 
         $rocket->mergePayload([
             'method' => 'alipay.fund.account.query',