+++ src/Plugin/Alipay/Fund/TransPagePayPlugin.php
@@ -17,7 +17,7 @@ class TransPagePayPlugin implements PluginInterface
 {
     public function assembly(Rocket $rocket, Closure $next): Rocket
     {
-        Logger::info('[alipay][TransPagePayPlugin] 插件开始装载', ['rocket' => $rocket]);
+        Logger::debug('[alipay][TransPagePayPlugin] 插件开始装载', ['rocket' => $rocket]);
 
         $rocket->setDirection(ResponseParser::class)
             ->mergePayload([